requirements.txt
exceptiongroup==1.1.1
Faker==18.4.0
greenlet==2.0.1
iniconfig==2.0.0
packaging==23.1
playwright==1.32.1
pluggy==1.0.0
pyee==9.0.4
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
